backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9
google-genai>=1.0.0
pydantic==2.9.2
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
PyJWT>=2.8.0
# clipspy==1.0.3  # Optional: requires C compiler. Fallback simulation is used if not installed.
